fixed stopping jenkins after test run

+++ Rakefile
@@ -10,7 +10,7 @@ end
 
 pidfile = 'jenkins/pidfile'
 
-task :start_jenkins => :stop_jenkins do
+task :start_jenkins do
   jenkins_pid = spawn 'java -DJENKINS_HOME=./jenkins/home -jar jenkins/jenkins.war'
   File.open(pidfile, 'w+') { |f| f.write(jenkins_pid) }
   puts "[JUNKIE] Jenkins starting... pid=#{jenkins_pid}"
@@ -19,10 +19,15 @@ task :start_jenkins => :stop_jenkins do
 end
 
 task :stop_jenkins do
-  puts "stopping"
   jenkins_pid = IO.read(pidfile) if File.exist?(pidfile)
-  puts "[JUNKIE] Killing pid=#{jenkins_pid}"
-  `kill -9 #{jenkins_pid}` if jenkins_pid
-  File.open(pidfile, 'w') { |f| f.write('') }
+  if jenkins_pid.nil? or jenkins_pid.empty?
+    puts '[JUNKIE] No pid found, unable to stop jenkins. Are you sure it is running?'
+    puts '[JUNKIE] If it is in the list below, please kill it manually:'
+    exec 'ps ax | grep jenkins'
+  else
+    puts "[JUNKIE] Killing pid=#{jenkins_pid}"
+    `kill -9 #{jenkins_pid}`
+    File.open(pidfile, 'w') { |f| f.write('') }
+  end
 end
